fix v4-in-v6-mapped handling, utilize stdlib for parsing
Since this library was proted from the corresponding ruby gem, it
also inherited some of its issues.
There is a long standing open issue regarding the handling of
v4-in-v6-mapped addresses:
https://github.com/ipaddress-gem/ipaddress/issues/98

The stdlib `Socket::IPAddress` parsing and formatting
(`#to_s`/`#address`) code already deals with it properly as per RFC5952.

By utilizing these public stdlib methods for the heavy lifting, we can
get rid of some kludges in this library.

Furthermore, it provides some nice efficiency benefits over the previous
implementation:
TODO:
  - Benchmark results

I'm unable to see the benefit of having a special `IPv6::Mapped`
subclass, that lugs around a nested `IPv4` instance.
In my opinion they should be treated as "normal" IPv6 addresses, since
everything is *supposed* to be compatible with it.
I have provided a `#IPv6.ipv4` helper to initialize a full `IPv4`
instance from a v4-in-v6-mapped address for situation that require it.

Be advised, this commit introduces breaking changes!

TODO:
  - benchmarks
  - adjust specs
  - adjust comments/docs